Cast the difference of two pointers to an int so that it can be compiled on more esoteric architectures.

+++ usr.sbin/pppd/pppdump/pppdump.c
@@ -315,7 +315,7 @@ dumpppp(f)
 			++r;
 			if (endp - r > mru)
 			    printf("     ERROR: length (%d) > MRU (%d)\n",
-				   endp - r, mru);
+				   (int)(endp - r), mru);
 			if (decompress && fcs == PPP_GOODFCS) {
 			    /* See if this is a CCP or compressed packet */
 			    d = dbuf;
